+++ backend/routes-loader/api.ts
@@ -4,7 +4,7 @@ import dataRoutes from "../routes/data.js"
 import emailsRoutes from "../routes/emails.js"
 import followupsRoutes from "../routes/followups.js"
 import franceConnectRoutes from "../routes/france-connect.js"
-import githubRoutes from "../routes/github.js"
+import moncompteproRoutes from "../routes/moncomptepro.js"
 import openfiscaRoutes from "../routes/openfisca.js"
 import outilsRoutes from "../routes/outils.js"
 import proxyRoutes from "../routes/proxy.js"
@@ -20,7 +20,7 @@ dataRoutes(api)
 emailsRoutes(api)
 followupsRoutes(api)
 franceConnectRoutes(api)
-githubRoutes(api)
+moncompteproRoutes(api)
 legacySmsRoutes(api)
 openfiscaRoutes(api)
 outilsRoutes(api)
